requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
jinja2==3.1.3
mysql-connector-python==8.3.0
pandas==2.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
unidecode==1.3.7
pydantic==2.7.0
